# backend/elrincondeeva/elrincondeeva/src/main/java/es/uv/hemal/elrincondeeva/domain/Role.java
package es.uv.hemal.elrincondeeva.domain;

import org.springframework.data.annotation.Id;
import org.springframework.data.relational.core.mapping.Column;
import org.springframework.data.relational.core.mapping.Table;

import java.util.List;

@Table("roles")
public class Role {
	@Id
	private Integer id;

	@Column("name")
	private String name;



	public Integer getId() {
		return id;
	}

	public void setId(Integer id) {
		this.id = id;
	}

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

}